requirements.txt
click
flask
flask-login
flask-sqlalchemy
pyyaml
